Update 5 done
Features

Digging is now a thing in Endless Mode! You can dig up a regular Treasure Chest which will award you with either a full Relic, a weapon or a lot of Bounty.

Gameplay

Super Shield now explodes when duration is over, dealing damage to enemies and pushing them away